Mentioned Microsoft Visual C++ Build Tools 2015
This is an installer for the C++ build tools that does not require a full installation of Visual Studio and provides the necessary linker. I think many people would be happy to know that there is a lighter option like this out there. The download link is on the same page, under Tools for Visual Studio 2015.
More info here: https://blogs.msdn.microsoft.com/vcblog/2016/03/31/announcing-the-official-release-of-the-visual-c-build-tools-2015/